refactor(stella-tui): the last three v1 modules are gone
`views/` held nothing that still rendered the pre-SPEC-5 way. What was left
was a shim, an orchestrator whose every pane was already v2, and a primitive
library the v2 tree imported eleven times.

- `views/graph.rs` was an empty state and a cursor clamp wrapping a call to
  `v2::graph_tab`. Both now sit in `graph_tab` with the drawing they guard,
  and its inner draw is `paint`; the tests that pinned the tab through the
  deck's frozen signature move with them, unchanged in what they assert.
- `views/settings.rs` dispatched all three panes into `v2::engine_panel`,
  `v2::tools` and `v2::seats`. It moves as it stands.
- `views/cards.rs` is the floating-card chrome the v2 tree already depended
  on, and `spinner_glyph` joins `status_glyph` in `theme`, which is the
  module that exists so every view pulls its glyphs from one place.

`keymap.rs`'s key-witness `include_str!` followed the settings move. It is
`#[cfg(test)]`, so `cargo check --lib` was green while `cargo test` could not
build — worth naming, because the same shape hides any path-coupled reference
from a check that only compiles the library.
