+++ crates/stella-tui/src/keymap.rs
@@ -645,7 +645,7 @@ mod tests {
     /// `handle_deck_key`. The rest are the four places a row's vocabulary
     /// genuinely lives elsewhere: `deck_ui/list_nav.rs` (the tree keys are
     /// that helper), `deck_ui/dispatch.rs` (the `>` steer marker),
-    /// `views/settings.rs` + `v2/engine_panel/keys.rs` (the SETTINGS panes and
+    /// `v2/settings.rs` + `v2/engine_panel/keys.rs` (the SETTINGS panes and
     /// their modal editor), `v2/subagents.rs` (the overlay's own verbs), and
     /// `deck_shell.rs` (`⌃V`, claimed by the run loop above the pure key
     /// layer because the capture is blocking I/O).
@@ -679,7 +679,7 @@ mod tests {
             include_str!("v2/engine_panel/keys.rs"),
             include_str!("v2/agents_page.rs"),
             include_str!("v2/subagents.rs"),
-            include_str!("views/settings.rs"),
+            include_str!("v2/settings.rs"),
         ]
     }
