+++ crates/stella-tui/src/v2/cards.rs
