+++ crates/stella-tui/src/v2/settings.rs
