+++ crates/stella-tui/src/v2/skills/overlays.rs
@@ -12,7 +12,7 @@
 //! theirs, so the body carries only what the dialog is about.
 //!
 //! The two pickers here select the way the tab body does and the way
-//! `crate::views::cards` states it: a `▸` marker glyph **and** a
+//! `crate::v2::cards` states it: a `▸` marker glyph **and** a
 //! [`token::HL`] background together, because the golden suite strips style —
 //! a tint with no glyph is invisible to it, and a glyph with no
 //! tint is a dialog row that reads differently from the identical row two boxes
@@ -109,7 +109,7 @@ pub fn render_prompt(ui: &DeckUi, now_ms: u64, area: Rect, buf: &mut Buffer) {
             );
         }
         Some(SkillPrompt::Creating { description, scope }) => {
-            let spinner = crate::views::spinner_glyph(now_ms, ui.no_anim);
+            let spinner = crate::theme::spinner_glyph(now_ms, ui.no_anim);
             let lines = vec![
                 Line::from(vec![
                     Span::styled(
